requirements.txt
# Pacotes principais de modelagem e treino
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
tensorflow==2.15.0

# Para salvar e carregar modelos
cloudpickle==3.0.0

# Para o app web
streamlit==1.32.2
python-dotenv==1.0.1
openai==1.10.0
httpx==0.25.2

# Para Jupyter Notebook
notebook==7.1.1
